Ajustes no botào de share
